rename `zk_token_elgamal` module to `pod`
