+++ zk-token-sdk/src/instruction/transfer.rs
@@ -24,7 +24,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        zk_token_elgamal::pod,
+        pod,
     },
     bytemuck::{Pod, Zeroable},
 };
