+++ zk-token-sdk/src/instruction/transfer_with_fee.rs
@@ -29,7 +29,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        zk_token_elgamal::pod,
+        pod,
     },
     bytemuck::{Pod, Zeroable},
 };
