+++ zk-token-sdk/src/instruction/withdraw.rs
@@ -16,7 +16,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        zk_token_elgamal::pod,
+        pod,
     },
     bytemuck::{Pod, Zeroable},
 };
