+++ zk-token-sdk/src/instruction/zero_balance.rs
@@ -18,7 +18,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        zk_token_elgamal::pod,
+        pod,
     },
     bytemuck::{Pod, Zeroable},
 };
